requirements.txt
discord.py
async-timeout